requirements.txt
# pytest
# pylint
# ipdb
# jupyterlab
# numpy
# pandas
# matplotlib
# seaborn
# scikit-learn
# db-dtypes
# google-cloud-bigquery
# google-cloud-bigquery-storage


streamlit
openai
pillow
pandas
matplotlib
python-dotenv
